+++ RELEASENOTES.md
@@ -156,6 +156,8 @@
 *   HLS:
     *   Add support for upstream discard including cancelation of ongoing load
         ([#6322](https://github.com/google/ExoPlayer/issues/6322)).
+*   Ogg: Allow non-contiguous pages
+    ([#7230](https://github.com/google/ExoPlayer/issues/7230)).
 *   Extractors:
     *   Add `IndexSeeker` for accurate seeks in VBR MP3 streams
         ([#6787](https://github.com/google/ExoPlayer/issues/6787)). This seeker